# Update aws_s3_bucket table to use region from list tables API Closes #2461