--- a/aws/table_aws_s3_object_version.go
+++ b/aws/table_aws_s3_object_version.go
@@ -110,7 +110,12 @@ func tableAwsS3ObjectVersion(_ context.Context) *plugin.Table {
 
 func listS3ObjectVersions(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	bucketName := d.EqualsQuals["bucket_name"].GetStringValue()
-	bucketRegion := h.HydrateResults["getBucketRegionForObjects"].(string)
+	bucketRegion, err := doGetBucketRegion(ctx, d, h, bucketName)
+	if err != nil {
+		return nil, err
+	} else if bucketRegion == "" {
+		return nil, nil
+	}
 
 	svc, err := S3Client(ctx, d, bucketRegion)
 	if err != nil {