# Integration test logs
<details>
  <summary>Logs</summary>

```
Add passing integration test logs here
```
</details>

# Example query results
<details>
  <summary>Results</summary>

```
> select * from aws_s3_bucket;
+-----------------------------------------------------------------+------------------------------------------------------------------------------+---------------------------+-------------------------+--------------------+-----------------------+-------------------+---------------------+-------------->
| name                                                            | arn                                                                          | creation_date             | bucket_policy_is_public | versioning_enabled | versioning_mfa_delete | block_public_acls | block_public_policy | ignore_public>
+-----------------------------------------------------------------+------------------------------------------------------------------------------+---------------------------+-------------------------+--------------------+-----------------------+-------------------+---------------------+-------------->
| cf-templates-yyk2l1c0d6k5-ap-south-1                            | arn:aws:s3:::cf-templates-yyk2l1c0d6k5-ap-south-1                            | 2025-02-11T17:15:52+05:30 | false                   | false              | false                 | true              | true                | true         >
| elasticbeanstalk-ap-south-1-xxxxxxxxxxxx                        | arn:aws:s3:::elasticbeanstalk-ap-south-1-xxxxxxxxxxxx                        | 2024-11-16T05:23:04+05:30 | false                   | false              | false                 | false             | false               | false        >
| my-waf-logs-bucket-test                                         | arn:aws:s3:::my-waf-logs-bucket-test                                         | 2024-11-21T03:18:11+05:30 | false                   | false              | false                 | true              | true                | true         >
| aws-security-data-lake-eu-central-1-5kzifb2pzuzw7kgalmmvfsh9pxq | arn:aws:s3:::aws-security-data-lake-eu-central-1-5kzifb2pzuzw7kgalmmvfsh9pxq | 2024-11-16T05:17:22+05:30 | false                   | false              | false                 | true              | true                | true         >
| aws-security-data-lake-eu-west-1-e1mtg9v0rolvlglniawrhjf4ialrti | arn:aws:s3:::aws-security-data-lake-eu-west-1-e1mtg9v0rolvlglniawrhjf4ialrti | 2024-11-16T05:17:22+05:30 | false                   | false              | false                 | true              | true                | true         >
| aws-security-data-lake-ap-northeast-1-yd5aigzmgod7uadzgou2d2otb | arn:aws:s3:::aws-security-data-lake-ap-northeast-1-yd5aigzmgod7uadzgou2d2otb | 2024-11-16T05:17:21+05:30 | false                   | false              | false                 | true              | true                | true         >
| aws-security-data-lake-ap-southeast-2-4i3pnujz7bpnx72x6oheacrcx | arn:aws:s3:::aws-security-data-lake-ap-southeast-2-4i3pnujz7bpnx72x6oheacrcx | 2024-11-16T05:17:22+05:30 | false                   | false              | false                 | true              | true                | true         >
| delete-me32                                                     | arn:aws:s3:::delete-me32                                                     | 2024-11-07T18:59:49+05:30 | true                    | true               | false                 | false             | false               | false        >
| vpc-subnet-flow-log-tp                                          | arn:aws:s3:::vpc-subnet-flow-log-tp                                          | 2025-02-25T22:18:52+05:30 | false                   | false              | false                 | true              | true                | true         >
| flowpipe-test-8a6db914-9a98-f789-e9d2-d13df22528ee              | arn:aws:s3:::flowpipe-test-8a6db914-9a98-f789-e9d2-d13df22528ee              | 2024-10-09T11:32:36+05:30 | false                   | false              | false                 | true              | true                | true         >
| network-firewall-test-tailpipe                                  | arn:aws:s3:::network-firewall-test-tailpipe                                  | 2025-03-03T20:05:02+05:30 | false                   | false              | false                 | true              | true                | true         >
| do-not-delete-rds-custom-xxxxxxxxxxxx-us-east-2-example         | arn:aws:s3:::do-not-delete-rds-custom-xxxxxxxxxxxx-us-east-2-example         | 2024-02-16T12:02:10+05:30 | false                   | false              | false                 | false             | false               | false        >
| flowpipe-test-56bac3ad-609e-73d6-77f1-dbea94d7b819              | arn:aws:s3:::flowpipe-test-56bac3ad-609e-73d6-77f1-dbea94d7b819              | 2024-10-09T11:34:20+05:30 | false                   | false              | false                 | true              | true                | true         >
| test-cloudtrail-s3-bucket-3d364669-8ca1-50d5-9e1b-db8ba3fd4289  | arn:aws:s3:::test-cloudtrail-s3-bucket-3d364669-8ca1-50d5-9e1b-db8ba3fd4289  | 2024-10-02T02:55:57+05:30 | false                   | false              | false                 | true              | true                | true         >
| aws-security-data-lake-us-west-2-xnssh4doirrzwpxmcfxy3vrbsbbu4k | arn:aws:s3:::aws-security-data-lake-us-west-2-xnssh4doirrzwpxmcfxy3vrbsbbu4k | 2024-11-12T14:13:15+05:30 | false                   | false              | false                 | true              | true                | true         >
| config-bucket-xxxxxxxxxxxx                                      | arn:aws:s3:::config-bucket-xxxxxxxxxxxx                                      | 2024-10-16T08:47:03+05:30 | false                   | false              | false                 | false             | false               | false        >
| capcha7.2                                                       | arn:aws:s3:::capcha7.2                                                       | 2024-11-15T02:24:12+05:30 | false                   | true               | false                 | true              | true                | true         >
| emr-studio-test-bucket-xxxxxxxxxxxx                             | arn:aws:s3:::emr-studio-test-bucket-xxxxxxxxxxxx                             | 2025-05-13T19:28:48+05:30 | false                   | false              | false                 | true              | true                | true         >
| mwaa-test-environment-1745905988                                | arn:aws:s3:::mwaa-test-environment-1745905988                                | 2025-04-29T11:23:12+05:30 | false                   | true               | false                 | true              | true                | true         >
| cost-optimization-recommendations-tp                            | arn:aws:s3:::cost-optimization-recommendations-tp                            | 2025-03-18T16:55:12+05:30 | false                   | false              | false                 | true              | true                | true         >
| fp-2024-10-04-04-52-48                                          | arn:aws:s3:::fp-2024-10-04-04-52-48                                          | 2024-10-04T10:22:59+05:30 | false                   | true               | false                 | true              | true                | true         >
| nlb-test-tailpipe                                               | arn:aws:s3:::nlb-test-tailpipe                                               | 2025-02-20T16:18:32+05:30 | false                   | false              | false                 | true              | true                | true         >
| cost-export-create-new                                          | arn:aws:s3:::cost-export-create-new                                          | 2025-03-27T13:03:39+05:30 | false                   | true               | false                 | true              | true                | true         >
| firewall-logs-1741250675                                        | arn:aws:s3:::firewall-logs-1741250675                                        | 2025-03-06T14:15:28+05:30 | false                   | true               | false                 | true              | true                | true         >
| clb-test-pc                                                     | arn:aws:s3:::clb-test-pc                                                     | 2025-02-17T21:09:57+05:30 | false                   | false              | false                 | true              | true                | true         >
| elasticbeanstalk-us-east-2-xxxxxxxxxxxx                         | arn:aws:s3:::elasticbeanstalk-us-east-2-xxxxxxxxxxxx                         | 2024-11-15T02:28:35+05:30 | false                   | false              | false                 | false             | false               | false        >
| test-s3-server-access-logs-ved                                  | arn:aws:s3:::test-s3-server-access-logs-ved                                  | 2025-04-08T01:35:12+05:30 | false                   | true               | false                 | true              | true                | true         >
| partha-test-finding-531                                         | arn:aws:s3:::partha-test-finding-531                                         | 2025-05-07T13:14:26+05:30 | false                   | false              | false                 | true              | true                | true         >
|                                                                 |                                                                              |                           |                         |                    |                       |                   |                     |              >
| aws-s3-tag3                                                     | arn:aws:s3:::aws-s3-tag3                                                     | 2024-03-01T14:29:21+05:30 | false                   | false              | false                 | true              | true                | true         >
| cost-usage-report-log                                           | arn:aws:s3:::cost-usage-report-log                                           | 2025-02-28T16:57:01+05:30 | false                   | false              | false                 | true              | true                | true         >
| test-s3-bucket-2600a5ba-da1e-d998-32e4-d20fa84662b1             | arn:aws:s3:::test-s3-bucket-2600a5ba-da1e-d998-32e4-d20fa84662b1             | 2024-10-01T21:54:56+05:30 | false                   | false              | false                 | true              | true                | true         >
| non-date-based-destination                                      | arn:aws:s3:::non-date-based-destination                                      | 2025-02-06T14:28:59+05:30 | false                   | true               | false                 | true              | true                | true         >
| aws-cloudtrail-logs-xxxxxxxxxxxx-427ec135                       | arn:aws:s3:::aws-cloudtrail-logs-xxxxxxxxxxxx-427ec135                       | 2024-09-24T19:57:48+05:30 | false                   | false              | false                 | true              | true                | true         >
|                                                                 |                                                                              |                           |                         |                    |                       |                   |                     |              >
| lambda-log-bucket-xxxxxxxxxxxx-us-east-1-test-lambda-log        | arn:aws:s3:::lambda-log-bucket-xxxxxxxxxxxx-us-east-1-test-lambda-log        | 2025-04-14T13:07:04+05:30 | false                   | true               | false                 | true              | true                | true         >
| test-cloudtrail-s3-bucket-858e1c38-38fd-981b-200a-ffc4afa25c1b  | arn:aws:s3:::test-cloudtrail-s3-bucket-858e1c38-38fd-981b-200a-ffc4afa25c1b  | 2024-10-04T00:19:33+05:30 | false                   | false              | false                 | true              | true                | true         >
| tp-securityhub-finding                                          | arn:aws:s3:::tp-securityhub-finding                                          | 2024-11-26T19:25:43+05:30 | false                   | false              | false                 | false             | false               | false        >
| test-vpc-flow-log-tp                                            | arn:aws:s3:::test-vpc-flow-log-tp                                            | 2025-02-12T19:50:05+05:30 | false                   | false              | false                 | true              | true                | true         >
| tes-encryption-31                                               | arn:aws:s3:::tes-encryption-31                                               | 2025-02-17T10:18:06+05:30 | false                   | false              | false                 | true              | true                | true         >
| cf-templates-yyk2l1c0d6k5-us-east-2                             | arn:aws:s3:::cf-templates-yyk2l1c0d6k5-us-east-2                             | 2024-10-16T08:09:43+05:30 | false                   | false              | false                 | true              | true                | true         >
| cost-usage-report-log-2-0-exclude-split-cost                    | arn:aws:s3:::cost-usage-report-log-2-0-exclude-split-cost                    | 2025-03-03T21:51:04+05:30 | false                   | false              | false                 | true              | true                | true         >
| awswafs3logs                                                    | arn:aws:s3:::awswafs3logs                                                    | 2024-11-06T10:36:01+05:30 | false                   | false              | false                 | true              | true                | true         >
| do-not-delete-ssm-diagnosis-xxxxxxxxxxxx-us-east-1-l07up        | arn:aws:s3:::do-not-delete-ssm-diagnosis-xxxxxxxxxxxx-us-east-1-l07up        | 2025-03-05T16:02:49+05:30 | false                   | true               | false                 | true              | true                | true         >
| fp-2024-10-04-04-55-20                                          | arn:aws:s3:::fp-2024-10-04-04-55-20                                          | 2024-10-04T10:25:24+05:30 | false                   | false              | false                 | true              | true                | true         >
| test-cloudtrail-s3-bucket-b1354b5e-93b0-8ebe-0d57-81ac9610bb9d  | arn:aws:s3:::test-cloudtrail-s3-bucket-b1354b5e-93b0-8ebe-0d57-81ac9610bb9d  | 2024-10-04T01:02:42+05:30 | false                   | true               | false                 | true              | true                | true         >
| date-based-partition-source-bucket-pc                           | arn:aws:s3:::date-based-partition-source-bucket-pc                           | 2025-02-06T11:47:00+05:30 | false                   | false              | false                 | true              | true                | true         >
| aws-waf-logs-test-s3-bucket                                     | arn:aws:s3:::aws-waf-logs-test-s3-bucket                                     | 2024-11-06T10:45:18+05:30 | false                   | false              | false                 | true              | true                | true         >
| test-tailpipe-source-pc                                         | arn:aws:s3:::test-tailpipe-source-pc                                         | 2025-02-04T16:44:50+05:30 | false                   | false              | false                 | true              | true                | true         >
| fp-2024-10-04-04-48-22                                          | arn:aws:s3:::fp-2024-10-04-04-48-22                                          | 2024-10-04T10:18:26+05:30 | false                   | true               | false                 | true              | true                | true         >
| aws-cloudtrail-logs-xxxxxxxxxxxx-3952022e                       | arn:aws:s3:::aws-cloudtrail-logs-xxxxxxxxxxxx-3952022e                       | 2025-05-12T11:21:30+05:30 | false                   | false              | false                 | true              | true                | true         >
112 rows
```

</details>