@ParthaI 2025-06-09
_Note: This PR requires regression testing since we're upgrading the AWS SDK S3 package version. Because this package is used by other tables as well, ensure there are no breaking changes._

Observation/Impact with the changes in this PR on the table `aws_s3_bucket`

Metric | Before Changes | After Changes | Difference
-- | -- | -- | --
Number of Buckets | 112 | 112 | ✅ No mismatch
Query Timing | 25.6s | 20.6s | ⬇️ Improved by 5.0s
Hydrate Calls | 1,904 | 1,792 | ⬇️ Reduced by 112 calls
Result Summary | Time: 25.6s. Rows returned: 112. Rows fetched: 112. Hydrate calls: 1,904.Scans:  1) aws_s3_bucket.aws_140: Time: 25.5s. Fetched: 112. Hydrates: 1,904. | Time: 20.6s. Rows returned: 112. Rows fetched: 112. Hydrate calls: 1,792.Scans:  1) aws_s3_bucket.aws: Time: 20.5s. Fetched: 112. Hydrates: 1,792. | ✅ Hydrate calls reduced


The changes maintain accuracy (no mismatch in bucket counts) while improving query performance and reducing the number of hydrate calls.