@ParthaI 2025-06-09
This PR is ready for review. I've tested all the `aws_s3_*` tables and applied necessary changes to handle region properly. I've compared the results between the old and new plugin versions and haven't encountered any breaking changes.